LayoutTests/platform/playstation-ps4/fast/ruby/ruby-justification-hittest-expected.txt

This test makes sure that hit testing works with ruby inside text-align: justify.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS right is 792
FAIL document.elementFromPoint(right - 1, bottom - 1).id should be ruby. Was .
PASS successfullyParsed is true

TEST COMPLETE
abcdefg abcdefga mmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmmm
